requirements.txt
torch>=1.12.0
torchvision>=0.13.0
timm>=0.6.7
numpy
Pillow
matplotlib
onnx
onnxruntime
onnxruntime-tools
pytorch-grad-cam
